Temperature is now a float, reduced max tokens

+++ StardewGPT/GptApiObjects.cs
@@ -32,7 +32,7 @@ namespace StardewGPT
     {
         public string model { get; set; }
         public string prompt { get; set; }
-        public int temperature { get; set; }
+        public float temperature { get; set; }
         public int max_tokens { get; set; }
         public string stop { get; set; }
     }
